pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devsec-scanner"
version = "0.1.0"
description = "AI-powered security scanner that finds actually exploitable vulnerabilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "colorama>=0.4.6",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "firebase-admin>=6.4.0",
    "gitpython>=3.1.40",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "rich>=13.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0"
]

[project.scripts]
devsec = "devsec_scanner.main:main"
